Fixed HTML generated with UTF-8 characters.
Without this encoding line, the generated HTML has an ASCII encoding
which, when non-ASCII characters are injected (e.g., from the gem
summary, description, etc.), will cause the write to disk to fail with
an encoding conversion error:

    Encoding::InvalidByteSequenceError: "\xC3" on US-ASCII

Regrettably, we were unable to construct a test that failed. We suspect
this is because the test suite setup is somehow causing all strings to
be encoded as UTF-8, either premptively or forced at a later
stage.