Backfill tests around whether dependencies are re-saved
We try to prevent the HTML/markdown reports from being updated with a
new timestamp every time you run license_finder.  The timestamp is based
on when the DB was last changed (a questionable approach... see bug
59043220) so we try to avoid changing the DB unnecessarily. These tests
assert that when a dependency's license isn't changing, the DB won't be
updated.