Stop looking at bundled dependencies for node.js
The `bundleDependencies` and `bundledDependencies` keys in npm’s
JSON output are not correct places to look for additional dependencies.

They are simple arrays of package names, instead of metadata about those
packages. And it seems that they are essentially a whitelist of packages
that should already be listed in `dependencies` or `devDependencies`.

https://www.npmjs.org/doc/files/package.json.html#bundleddependencies