Introducing a Logger class
which is controlled by CLI options --quiet and --debug.

The goal is to make more transparent the things that LicenseFinder is
doing. With this commit, that's only which package managers are active.
But in the future I'd like to also include which packages are inspected,
what their dependencies are, and how the license information is being
derived/discovered.

I've also introduced a shared spec for PackageManager classes.