Avoid platform-specific shell commands in package managers
Strictly speaking "cd" is a platform-specific shell command, although it
happens to be available on all major platforms under the same name. The
"&&" however is not: On Windows you would need to use a single "&"
instead. It just happenes to work as Windows is interpreting this as two
single "&" with an empty command in between.

While cleaning this up, also move the "capture" helper to the base class
and consistently use it in all package managers.