Fix tests to also not rely on shell commands
The basic idea always is to stub "chdir" to not throw an error as the fake
paths do not exist, but still call its original block. "capture", being
the return value of the block, is then stubbed to return fake data when
called with the expected package manager's commands to determine
dependencies.