Make the package managers the canonical place for their commands
Do not implement fallbacks for commands in the configuration *and* package
managers, but only in the latter for clarity. Remove the Gradle test that
checks for a default command in the configuration.

Also, do not repeat the package command as part of the full command line,
but simply insert the package management command.