Make yarn prepare_command use --ignore-engines flag
package.json can contain an "engines" key that can be used to specify
the runtime versions that the project shoulod run on.
e.g:
"engines": {
    "node": ">= 0.11.16",
    "npm" : "~1.0.20"
}

When `yarn install` is run, yarn checks that the local runtimes match
the specified versions.
`yarn install` will fail if a specified engin version is not found.

This behaviour is undesirable when looking for licenses. It can be
avoided by running yarn with the `--ignore-engines` flag.

This has been available since yarn 0.13.0, Sept 2016:
https://github.com/yarnpkg/yarn/commit/a0b8195c012eda24ce4ea8b9da60dce29794012a