Replace whitelist terminology
Replace whitelist terminology with the concept of permitted licenses.
The new terminology aims to be more intuitive and
descriptive of the functionality provided to users.

This is a breaking change and users with an existing
dependency decisions file will need to find and replace all instances
of `whitelist` with `permit`.

https://developers.google.com/style/word-list#blacklist
https://github.com/rubocop-hq/rubocop/pull/6464