Update Nuget Package Manager prepare command
Nuget failed to 'restore' projects when it only contained the
'packages.config' file. Nuget would complain that it did not know where
to download the packages to.

So, the Nuget package manager has been updated to try the original
command, and if that fails, it would try again but this time providing
an -PackagesDirectory which is set to the current directory.

[#173420224]

Signed-off-by: Jason Smith <jsmith@pivotal.io>