Change Go modules to list imported modules
If module mode is enabled (`-m`), all possible module dependencies are
listed, regardless of whether the packages requiring the module are
actually imported.

This results in un-imported modules being listed, which is incongruous
with other Go package manager implementations.

The removal of the `-m` flag causes `go list` to list imported packages.
A format string is used so that the corresponding module is listed in
place of each imported package.