Create licenses.json
This is a list of licenses in JSON format, to be used when packages being scanned do not provide license names, only URLs.  Is that case we can find license name by URL reference, if exists.  This is not a final/comprehensive list, but could be a good starting point.