Update the default timezone to GMT
During the Ubuntu Bionic upgrade, the default timezome was set to
America/Toronto in order to install the new version of Composer. The
newer version of composer requires a timezome to bet set in order to
install.

However, the test assumes LF runs on GMT timezome image.  So, the
default timezone has been updated.

[#178471230]