Fix CocoaPods plutil argument escaping
An attempt was made to fix a command injection vector in
https://github.com/pivotal/LicenseFinder/commit/b0a61a2d833921c714cc39cdda8ba80af3f33d04

Whitelisting specific characters that can be allowed in a path is prone to failures
(https://github.com/pivotal/LicenseFinder/issues/846), especially in
non-english locales.

Instead of trying to work around this by blocking usage of certain
characters, we can use one of Ruby's parameterized methods of command
execution which will properly handle shell escaping.