Safely load dependency decisions with any version of Psych
Psych 4 no longer loads YAML unsafely with Psych.load. This change
switches to Psych.safe_load. Since the calling convention for
Psych.safe_load was changed, and the old convention is no longer
supported in Psych 4, either the old or the new convention is used based
on the loaded version of Psych.