:wrench: allow to specify cocoadpods acknowledgment through ENV
 > in multi projects folder of iOS projects, it looks like the cocoapods plugin is unable
to find the proper Pods-...-acknowledgments.plist file. As it takes only the 1st item of the dir,
We should be able to also specify the one we want trought an environment variable ACKNOWLEDGEMENT_PATH
just like what it is done in the SPM plugin where we do need to specify the environment variables for
DerivedDatas & Package.resolved files.