Add ability to specify package version when setting a manual license
Currently, adding a license for a package applies it to all versions of
that package. However, packages sometimes change their license between
versions.

This change allows the --version option to be specified when invoking
`license_finder licenses [add | remove]`. Similar to how approvals work,
once licenses for specific versions of a given dependency are added,
any licenses added for that dependency without the --version option are
ignored.