[Fixed] Ensure licenses are recognised by `spdx_id`
The recognition of licenses named by their SPDX ID is currently
inconsistent; for some licenses the ID is included as an alternative
name, but for some it is not. This leads to inconsistent behaviour, as
seen in issue #984.

Changing the matching logic to check the SPDX ID will make this more
consistent and intuitive.

This fixes #984